requirements.txt
pandas>=2.2.0
openpyxl>=3.1.0
PyYAML>=6.0.0
chromadb>=0.5.0
google-genai>=0.3.0
google-generativeai>=0.8.0
streamlit>=1.44.0
slack-bolt>=1.21.0
python-dotenv>=1.0.0
